PPieces.h:  const- and null- correctness work
also stdint work

+++ GP/SelOPlay.cpp
@@ -649,7 +649,7 @@ BOOL CSelList::HasNonOwnedPieces() const
         if (pSel.m_pObj->GetType() == CDrawObj::drawPieceObj)
         {
             const CPieceObj& pPObj = static_cast<const CPieceObj&>(*pSel.m_pObj);
-            if (pPTbl->GetOwnerMask(pPObj.m_pid) == 0)
+            if (pPTbl->GetOwnerMask(pPObj.m_pid) == uint32_t(0))
                 return TRUE;
         }
     }
@@ -665,7 +665,7 @@ BOOL CSelList::HasOwnedPieces() const
         if (pSel.m_pObj->GetType() == CDrawObj::drawPieceObj)
         {
             const CPieceObj& pPObj = static_cast<const CPieceObj&>(*pSel.m_pObj);
-            if (pPTbl->GetOwnerMask(pPObj.m_pid) != 0)
+            if (pPTbl->GetOwnerMask(pPObj.m_pid) != uint32_t(0))
                 return TRUE;
         }
     }
@@ -793,8 +793,8 @@ void CSelList::LoadTableWithOwnerStatePieceIDs(std::vector<PieceID>& pTbl, LoadF
         if (pSel.m_pObj->GetType() == CDrawObj::drawPieceObj)
         {
             CPieceObj& pPObj = static_cast<CPieceObj&>(*pSel.m_pObj);
-            if (eWantOwned == LF_OWNED    && pPTbl->GetOwnerMask(pPObj.m_pid) != 0 ||
-                eWantOwned == LF_NOTOWNED && pPTbl->GetOwnerMask(pPObj.m_pid) == 0 ||
+            if (eWantOwned == LF_OWNED    && pPTbl->GetOwnerMask(pPObj.m_pid) != uint32_t(0) ||
+                eWantOwned == LF_NOTOWNED && pPTbl->GetOwnerMask(pPObj.m_pid) == uint32_t(0) ||
                 eWantOwned == LF_BOTH)
             {
                 pTbl.push_back(pPObj.m_pid);
